move check script, module, and declare_script_or_function to error checker
