+++ crates/generated_parser/src/ast_builder.rs
@@ -1,6 +1,5 @@
 use crate::context_stack::{
-    BindingInfo, BindingKind, BindingsIndex, ContextMetadata, ControlInfo, LabelIndex, LabelInfo,
-    LabelKind,
+    BindingInfo, BindingKind, BindingsIndex, ContextMetadata, ControlInfo, LabelInfo, LabelKind,
 };
 use crate::declaration_kind::DeclarationKind;
 use crate::early_error_checker::*;
@@ -4804,67 +4803,6 @@ impl<'alloc> AstBuilder<'alloc> {
         Ok(())
     }
 
-    // Declare bindings to script-or-function-like context, where function
-    // declarations are body-level.
-    fn declare_script_or_function<T>(
-        &self,
-        context: &mut T,
-        index: BindingsIndex,
-    ) -> Result<'alloc, ()>
-    where
-        T: LexicalEarlyErrorsContext + VarEarlyErrorsContext,
-    {
-        for info in self.context_metadata.bindings_from(index) {
-            match info.kind {
-                BindingKind::Var => {
-                    context.declare_var(
-                        info.name,
-                        DeclarationKind::Var,
-                        info.offset,
-                        &self.atoms.borrow(),
-                    )?;
-                }
-                BindingKind::Function | BindingKind::AsyncOrGenerator => {
-                    context.declare_var(
-                        info.name,
-                        DeclarationKind::BodyLevelFunction,
-                        info.offset,
-                        &self.atoms.borrow(),
-                    )?;
-                }
-                BindingKind::Let => {
-                    context.declare_lex(
-                        info.name,
-                        DeclarationKind::Let,
-                        info.offset,
-                        &self.atoms.borrow(),
-                    )?;
-                }
-                BindingKind::Const => {
-                    context.declare_lex(
-                        info.name,
-                        DeclarationKind::Const,
-                        info.offset,
-                        &self.atoms.borrow(),
-                    )?;
-                }
-                BindingKind::Class => {
-                    context.declare_lex(
-                        info.name,
-                        DeclarationKind::Class,
-                        info.offset,
-                        &self.atoms.borrow(),
-                    )?;
-                }
-                _ => {
-                    panic!("Unexpected binding found {:?}", info);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
     // Check bindings in function with FormalParameters.
     fn check_function_bindings(
         &mut self,
@@ -4930,36 +4868,6 @@ impl<'alloc> AstBuilder<'alloc> {
         Ok(())
     }
 
-    // Check bindings in Script.
-    fn check_script_bindings(&mut self) -> Result<'alloc, ()> {
-        let mut context = ScriptEarlyErrorsContext::new();
-        let index = BindingsIndex { index: 0 };
-        self.declare_script_or_function(&mut context, index)?;
-        self.context_metadata.pop_bindings_from(index);
-
-        let label_index = LabelIndex { index: 0 };
-        self.context_metadata.pop_labels_from(label_index);
-
-        self.check_unhandled_break_or_continue(context, 0)?;
-
-        Ok(())
-    }
-
-    // Check bindings in Module.
-    fn check_module_bindings(&mut self) -> Result<'alloc, ()> {
-        let mut context = ModuleEarlyErrorsContext::new();
-        let index = BindingsIndex { index: 0 };
-        self.declare_script_or_function(&mut context, index)?;
-        self.context_metadata.pop_bindings_from(index);
-
-        let label_index = LabelIndex { index: 0 };
-        self.context_metadata.pop_labels_from(label_index);
-
-        self.check_unhandled_break_or_continue(context, 0)?;
-
-        Ok(())
-    }
-
     // Returns IsSimpleParameterList of `params`.
     //
     // NOTE: For Syntax-only parsing (NYI), the stack value for FormalParameters
@@ -5009,5 +4917,8 @@ impl<'alloc> EarlyErrorChecker<'alloc> for AstBuilder<'alloc> {
     }
     fn context_metadata_immutable(&self) -> &ContextMetadata {
         &self.context_metadata
+    }
+    fn atoms(&self) -> &Rc<RefCell<SourceAtomSet<'alloc>>> {
+        &self.atoms
     }
 }
